Minor CSS updates and fixes. (#102)
* Remove unused "optional" css.

"optional" css was added pre-emptively but has never been used. It might
be of benefit to a larger project, but isnt necessary on small static
sites such as this one.

* Match width of footer to width of page.

Header and content were previously 80% of window, whereas footer was
(100% - 40px). Now all three are 80% of window.

* Fix body tag position.

body tag should come immediately after head tag.

* Remove duplicated/invalid css.

* Ensure footer is always on window bottom.

This ensures the footer is not left floating in the middle of the window
in the cases where page content is not >= window height.